Undoing some patterns that I naturally use
Splitting the models into their own, smaller components, as this is how
most apps work.  It made it difficult to understand the difference
between MVC and MVVM

Switched the two MVC Views to use this.publicFunction and to construct
their own commands.  That it's valid to create the commands in the View
is an important difference between MVC and the MVP/MVVM patterns